+++ tests/pyutilkit/test_files.py
@@ -1,0 +1,21 @@
+import pytest
+
+from pyutilkit.files import handle_exceptions
+
+
+def test_handle_exceptions_handled_exception() -> None:
+    @handle_exceptions(exceptions=(ZeroDivisionError,), default=0.0)
+    def invert(n: int) -> float:
+        return 1 / n
+
+    assert invert(1) == 1
+    assert invert(0) == 0
+
+
+def test_handle_exceptions_unhandled_exception() -> None:
+    @handle_exceptions(exceptions=(TypeError,), default=0.0)
+    def invert(n: int) -> float:
+        return 1 / n
+
+    assert invert(1) == 1
+    pytest.raises(ZeroDivisionError, invert, 0)